simple inventory + collectables
Inventory can be opened by pressing `i`
The inventory stores items from the 1st slot to the last in order of how they were picked up. If an item already exists, its added to that items count, not a new slot. When hovering over an item slot, it displays its name.

Added two types of gems, red and blue, as per the stories demmands.

As of yet, the inventory has no real use